requirements.txt
# Backend Prototype - requirements.txt
# List of Python packages required for the backend application.

Flask>=2.0
Flask-SQLAlchemy>=2.5
Flask-Migrate>=3.0
python-dotenv>=0.19
Flask-Bcrypt>=1.0
PyJWT>=2.0 # For JWT token handling
psycopg2-binary # PostgreSQL adapter (use appropriate adapter for your DB)

